posting_api/requirements/common.txt
asyncpg==0.21.0
databases==0.4.1
fastapi==0.61.1
lxml==4.6.2
pydantic==1.7.3
python-decouple==3.3
psycopg2-binary==2.8.6
requests==2.25.0
SQLAlchemy==1.3.20
uvicorn==0.12.3
urllib3==1.26.2
